+++ components/Alert.js
@@ -1,16 +1,16 @@
-import { FaInfoCircle } from "react-icons/fa";
-import { MdError, MdOutlineWarningAmber } from "react-icons/md";
-import { FaRegCheckCircle } from "react-icons/fa";
+import { ExclamationTriangleIcon } from "@heroicons/react/24/outline";
+import { ExclamationCircleIcon } from "@heroicons/react/24/outline";
+import { CheckCircleIcon } from "@heroicons/react/24/outline"
 
 export default function alert({ type, message }) {
   switch (type) {
     case "success":
       return (
         <div
-          className={`alert-${type} bg-green-100 rounded-lg py-5 px-6 mb-3 text-base text-green-700 inline-flex items-center w-full animate-pulse`}
+          className={`alert-${type} bg-green-100 rounded-lg py-5 px-6 mb-3 text-base text-green-700 inline-flex items-center w-full`}
           role="alert"
         >
-          <FaRegCheckCircle className="mr-2" />
+          <CheckCircleIcon className="h-5 w-5 mr-2" />
           <span>{message}</span>
         </div>
       );
@@ -18,10 +18,10 @@ export default function alert({ type, message }) {
     case "error":
       return (
         <div
-          className={`alert-${type} bg-red-100 rounded-lg py-5 px-6 mb-3 text-base text-red-700 inline-flex items-center w-full animate-pulse`}
+          className={`alert-${type} bg-red-100 rounded-lg py-5 px-6 mb-3 text-base text-red-700 inline-flex items-center w-full`}
           role="alert"
         >
-          <MdError className="mr-2" />
+          <ExclamationCircleIcon className="h-5 w-5 mr-2" />
           <span>{message}</span>
         </div>
       );
@@ -29,10 +29,10 @@ export default function alert({ type, message }) {
     case "warning":
       return (
         <div
-          className={`alert-${type} bg-yellow-100 rounded-lg py-5 px-6 mb-3 text-base text-yellow-700 inline-flex items-center w-full animate-pulse`}
+          className={`alert-${type} bg-yellow-100 rounded-lg py-5 px-6 mb-3 text-base text-yellow-700 inline-flex items-center w-full`}
           role="alert"
         >
-          <MdOutlineWarningAmber className="mr-2" />
+          <ExclamationTriangleIcon className="h-5 w-5 mr-2" />
           <span>{message}</span>
         </div>
       );
@@ -40,10 +40,10 @@ export default function alert({ type, message }) {
     case "info":
       return (
         <div
-          className={`alert-${type} bg-blue-100 rounded-lg py-5 px-6 mb-3 text-base text-blue-700 inline-flex items-center w-full animate-pulse`}
+          className={`alert-${type} bg-blue-100 rounded-lg py-5 px-6 mb-3 text-base text-blue-700 inline-flex items-center w-full`}
           role="alert"
         >
-          <FaInfoCircle className="mr-2" />
+          <ExclamationCircleIcon className="h-5 w-5 mr-2" />
           <span>{message}</span>
         </div>
       );