+++ pages/account/no-profile.js
@@ -2,13 +2,10 @@ import { authOptions } from "../api/auth/[...nextauth]";
 import { getServerSession } from "next-auth/next";
 import Page from "@components/Page";
 import PageHead from "@components/PageHead";
+import Link from "@components/Link";
 
 export async function getServerSideProps(context) {
-  const session = await getServerSession(
-    context.req,
-    context.res,
-    authOptions
-  );
+  const session = await getServerSession(context.req, context.res, authOptions);
 
   if (!session) {
     return {
@@ -43,18 +40,18 @@ export default function NoProfile() {
                 Please follow the QuickStart guide to create your Profile
               </p>
               <div className="mt-10 flex items-center justify-center gap-x-6">
-                <a
+                <Link
                   href="https://linkfree.eddiehub.io/eddiejaoude"
                   className="rounded-md bg-secondary-high px-3.5 py-1.5 text-base font-semibold leading-7 text-white shadow-sm hover:bg-secondary-low focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-secondary-high"
                 >
                   Example
-                </a>
-                <a
+                </Link>
+                <Link
                   href="https://linkfree.eddiehub.io/docs/quickstart"
                   className="text-base font-semibold leading-7 text-primary-high"
                 >
                   QuickStart Guide <span aria-hidden="true">→</span>
-                </a>
+                </Link>
               </div>
             </div>
           </div>