+++ pages/events.js
@@ -1,8 +1,10 @@
 import { useState } from "react";
 import { FaListUl, FaMicrophoneAlt } from "react-icons/fa";
 import { MdOutlineOnlinePrediction, MdOutlinePeople } from "react-icons/md";
+import { TbCoin, TbCoinOff } from "react-icons/tb";
 
-import logger from "@config/logger";
+import { getEvents } from "./api/events";
+
 import EventCard from "@components/event/EventCard";
 import Page from "@components/Page";
 import { EventTabs } from "@components/event/EventTabs";
@@ -10,34 +12,7 @@ import PageHead from "@components/PageHead";
 import Badge from "@components/Badge";
 
 export async function getServerSideProps(context) {
-  let events = [];
-  try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/events`);
-    events = await res.json();
-  } catch (e) {
-    logger.error(e, "ERROR events list");
-  }
-
-  // remove any invalid events
-  events = events.filter((event) => {
-    const dateTimeStyle = {
-      dateStyle: "full",
-      timeStyle: "long",
-    };
-    try {
-      new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-        new Date(event.date.start)
-      );
-      new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-        new Date(event.date.end)
-      );
-
-      return true;
-    } catch (e) {
-      logger.error(e, `ERROR event date for: "${event.name}"`);
-      return false;
-    }
-  });
+  let events = await getEvents();
 
   return {
     props: { events },
@@ -45,46 +20,59 @@ export async function getServerSideProps(context) {
 }
 
 export default function Events({ events }) {
-  let categorisedEvents = {
+  let categorizedEvents = {
     all: events,
     virtual: events.filter((event) => event.isVirtual === true),
     inPerson: events.filter((event) => event.isInPerson === true),
-    cfpOpen: events.filter((event) =>
-      event.date.cfpClose ? new Date(event.date.cfpClose) > new Date() : false
-    ),
+    cfpOpen: events.filter((event) => event.date.cfpOpen === true),
+    free: events.filter((event) => event.price?.startingFrom === 0),
+    paid: events.filter((event) => event.price?.startingFrom > 0),
   };
-  const filters = [
+  const tabFilters = [
     {
       title: "Show all",
       description: "List all events with no filters",
       key: "all",
       icon: FaListUl,
-      total: categorisedEvents.all.length,
+      total: categorizedEvents.all.length,
     },
     {
       title: "CFP open",
       description: "You can submit a talk to this conference",
       key: "cfpOpen",
       icon: FaMicrophoneAlt,
-      total: categorisedEvents.cfpOpen.length,
+      total: categorizedEvents.cfpOpen.length,
     },
     {
       title: "In person",
       description: "These are in person events",
       key: "inPerson",
       icon: MdOutlinePeople,
-      total: categorisedEvents.inPerson.length,
+      total: categorizedEvents.inPerson.length,
     },
     {
       title: "Virtual",
       description: "Held virtually online event",
       key: "virtual",
       icon: MdOutlineOnlinePrediction,
-      total: categorisedEvents.virtual.length,
+      total: categorizedEvents.virtual.length,
+    },
+    {
+      title: "Free",
+      description: "These events are free to attend",
+      key: "free",
+      icon: TbCoinOff,
+      total: categorizedEvents.free.length,
+    },
+    {
+      title: "Paid",
+      description: "These events are paid to attend",
+      key: "paid",
+      icon: TbCoin,
+      total: categorizedEvents.paid.length,
     },
   ];
 
-  const [tabs, setTabs] = useState(filters);
   const [eventType, setEventType] = useState("all");
 
   return (
@@ -106,19 +94,18 @@ export default function Events({ events }) {
           </Badge>
         </div>
         <EventTabs
-          tabs={tabs}
+          tabs={tabFilters}
           eventType={eventType}
           setEventType={setEventType}
         />
         <h2 className="text-md md:text-2xl text-lg text-primary-high font-bold md:mb-6 mb-3">
-            {filters.find((filter) => filter.key === eventType).description}
+          {tabFilters.find((filter) => filter.key === eventType).description}
         </h2>
         <ul role="list" className="divide-y divide-primary-low mt-6">
-
-          {categorisedEvents[eventType]?.map((event) => (
+          {categorizedEvents[eventType]?.map((event) => (
             <EventCard
               event={event}
-              username={event.username}
+              usernames={event.usernames}
               key={`${event.name} ${event.username}`}
             />
           ))}