+++ tests/setup/global-setup.js
@@ -1,4 +1,5 @@
 import fs from "fs";
+import { clientEnv } from "@config/schemas/clientSchema";
 
 const { USERS } = require("./test-users.js");
 
@@ -57,5 +58,17 @@ module.exports = async () => {
   } catch (e) {
     logger.error(e);
     throw new Error(`Test data "_test-wcag-user" not created`);
+  }
+
+  try {
+    const response = await fetch(
+      `${clientEnv.NEXT_PUBLIC_BASE_URL}/api/system/reload?secret=development`
+    );
+    if (response.status !== 200) {
+      throw new Error(`Test data not loaded into database`);
+    }
+  } catch (e) {
+    logger.error(e);
+    throw new Error(`Test data not loaded into database`);
   }
 };